Added header and some initial resources.
Added initial content mostly porting from the dev-doc repository,
which will be discontinued. All links of that repository will be
put into this one and various useful recipes and notes will be
moved to the skimmy-notes repository. In this way, links and dev notes
will be separated and more manageable.